+++ src/netcfg/netcfg-changetype.cpp
@@ -18,5 +18,5 @@
 
 NetCfgChangeType operator|(const NetCfgChangeType &a, const NetCfgChangeType &b)
 {
-    return NetCfgChangeType(static_cast<std::uint16_t>(a) | static_cast<std::uint16_t>(b));
+    return NetCfgChangeType(static_cast<std::uint32_t>(a) | static_cast<std::uint32_t>(b));
 }
